Makefile, platina-mk1.mk:  Removed _amd64 from names
The _amd64 suffix was preventing new views from building
the debian kernel.  The Makefile was looking in platina-mk1-amd64
for the config.  This worked on older views, since this directory
did in fact exist previously.  New views did not have this
extraneous directory and failed.  Removing amd64 from the names
corrected the issue.

Signed-off-by: Mark Pleso <mpleso@platinasystems.com>